hostezon project
hosteZON is developed help students to Get all Their Required Information at the same place. This is basically a combination of hostel + amazon which comes out as hosteZON. It has all the features like Blogs, Social, as well as an e-commerce servie is also included to make it more useful

we hope this Portal meet all your expectations and needs. Please share this to all of your friends so that they can also benefitted.
this is purely my project you can confirm it from my github account